config-editor: extract neighbor constraint validation and tile config validation againts coordinate into separate files
